Fixed an issue with datasources alert
In the event that replication for a cluster is switched off,
the datasources ls output won't include the "progress" attribute
that I was grepping on.  Now slurping in the entire ls output, and
in ruby (for ease of writing the more-complicated regex) am grepping
for the relevant lines by looking for datasource name and role.