Add master/slave monitoring & metrics to RedisMonitor
* reports the current 'role'
* if slave, then also reports:
  * 'master_link_status' (1 for up, 0 for down)
  * 'master_last_io_seconds_ago'
  * 'master_sync_in_progress'

The thought is you'd be able ot install this on a master/slave pair, and
it'd be collecting the correct metrics regardless of which is master and
which is slave.

In addition, reporting master_link_status as 1 or 0 allows
you to create a Plateau Trigger with low values, and alert aftering
after it has been down for N units of time.