Updates
Added a memory check to see the last status of the port. This way we will only send an alert if the port was up last run and is down this run.

Added a trigger to the yaml spec which will send an alert if there are no access requests being reported.